Collect definitions using both completion locations and definition locations
Completion locations for a given semantic node are treated as canonical,
while definitions are treated as non-canonical, unless the semantic node has
no completion, in which case we arbitrarily choose the first definition to
be the canonical one. Only canonical locations are added to the xref list,
with the canonical location being treated as the completion location.

This allows us to correctly handle cases involving templates with out-of-line
definitions where a definition has more than one semantic node and its
references and completions use different semantic nodes.